查询用户信息接口

+++ business/src/main/java/com/xingyi/logistic/authentication/bean/BaseUserRoles.java
@@ -3,23 +3,23 @@ package com.xingyi.logistic.authentication.bean;
 import com.xingyi.logistic.business.bean.BaseModelAndDO;
 
 public class BaseUserRoles extends BaseModelAndDO{
-    private Integer userId;
-    private Integer roleId;
+    private Long userId;
+    private Long roleId;
     private Integer isDeleted;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
